backend/community-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]